django/requirements.txt
Django==2.2.10
gunicorn==19.9.0
psycopg2==2.7.5
psycopg2-binary==2.7.5

django-bootstrap4==1.1.1
django-extensions==2.2.6
django-filter==2.2.0
djangorestframework==3.11.0


django-crispy-forms==1.7.2
pytz==2018.5
ptvsd==3.0.0
